move equipment dropdown to own component
