+++ src/app/treasure-hunt/treasure-chest/treasure-chest-menu/treasure-chest-menu.component.html
@@ -26,7 +26,7 @@
   <app-utility-type-dropdown [settings]="settings"></app-utility-type-dropdown>
   <!--filter by calculator type dropdown-->
   <app-calculator-type-dropdown [settings]="settings"></app-calculator-type-dropdown>
-  
+
   <!--manage cards button group-->
   <div class="nav-item" *ngIf="!inReport">
     <div class="btn-group">
@@ -42,25 +42,11 @@
     <button type="button" class="nav-btn" (click)="toggleAdditionalFilters()">Add'l Filters</button>
   </div>
 
-  <!--teams. use dropdown in report-->
+  <!--teams use dropdown in report-->
   <app-teams-dropdown *ngIf="inReport" [inReport]="inReport" [settings]="settings"></app-teams-dropdown>
 
   <!--equipment. use dropdown in report-->
-  <div *ngIf="inReport" class="nav-item dropdown">
-    <a class="nav-link dropdown-toggle" (click)="toggleEquipment()">
-      Equipment
-    </a>
-    <div class="dropdown-menu p-0 teams-dropdown" [ngClass]="{'show': displayEquipment == true}">
-      <div class="list-group">
-        <a class="list-group-item list-group-item-action" *ngFor="let option of equipments"
-          [ngClass]="{'active': option.selected}" (click)="setSelectedEquipment(option)">{{option.display}}
-          <span class="pull-right pl-2 pr-2" [ngClass]="{'badge badge-pill': option.numCalcs > 0}">
-            {{option.numCalcs}}
-          </span>
-        </a>
-      </div>
-    </div>
-  </div>
+  <app-equipments-dropdown *ngIf="inReport" [inReport]="inReport" [settings]="settings"></app-equipments-dropdown>
 
 </nav>
 <!--displays selected filter options-->
@@ -99,15 +85,7 @@
         <app-teams-dropdown [inReport]="inReport" [settings]="settings"></app-teams-dropdown>
       </div>
       <div class="col-6 pl-1">
-        <h5>Select Equipment</h5>
-        <div class="list-group list-group-flush">
-          <a class="list-group-item list-group-item-action" *ngFor="let equipment of equipments"
-            [ngClass]="{'active': equipment.selected}" (click)="setSelectedEquipment(equipment)">{{equipment.display}}
-            <span class="pull-right pl-2 pr-2" [ngClass]="{'badge badge-pill': equipment.numCalcs > 0}">
-              {{equipment.numCalcs}}
-            </span>
-          </a>
-        </div>
+        <app-equipments-dropdown [inReport]="inReport" [settings]="settings"></app-equipments-dropdown>
       </div>
     </div>
   </div>
