+++ src/app/treasure-hunt/treasure-chest/treasure-chest-menu/treasure-chest-menu.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input, ViewChild, ElementRef, HostListener } from '@angular/core';
-import { TreasureHunt, FilterOption } from '../../../shared/models/treasure-hunt';
+import { TreasureHunt } from '../../../shared/models/treasure-hunt';
 import { Settings } from '../../../shared/models/settings';
 import * as _ from 'lodash';
 import { Subscription } from 'rxjs';
@@ -7,9 +7,6 @@ import { TreasureChestMenuService } from './treasure-chest-menu.service';
 import { SortCardsData } from '../opportunity-cards/sort-cards-by.pipe';
 import { OpportunityCardsService, OpportunityCardData } from '../opportunity-cards/opportunity-cards.service';
 import { trigger, state, style, animate, transition } from '@angular/animations';
-import { processEquipmentOptions } from '../../calculators/opportunity-sheet/general-details-form/processEquipmentOptions';
-import { TreasureHuntService } from '../../treasure-hunt.service';
-import { SortCardsService } from '../opportunity-cards/sort-cards.service';
 
 @Component({
   selector: 'app-treasure-chest-menu',
@@ -40,10 +37,6 @@ export class TreasureChestMenuComponent implements OnInit {
     this.getNavbarWidth();
   }
 
-  // teams: Array<FilterOption>;
-  equipments: Array<FilterOption>;
-
-  displayEquipment: boolean = false;
   displayAdditionalFiltersDropdown: string = 'hide';
   sortByDropdown: boolean = false;
   treasureHunt: TreasureHunt;
@@ -57,21 +50,13 @@ export class TreasureChestMenuComponent implements OnInit {
   showImportModalSub: Subscription;
   showExportModal: boolean;
   showExportModalSub: Subscription;
-  constructor(private opportuntityCardsService: OpportunityCardsService, private treasureChestMenuService: TreasureChestMenuService,
-    private treasureHuntService: TreasureHuntService, private sortCardsService: SortCardsService) { }
+  constructor(private opportuntityCardsService: OpportunityCardsService, private treasureChestMenuService: TreasureChestMenuService) { }
 
   ngOnInit() {
     this.sortBySub = this.treasureChestMenuService.sortBy.subscribe(val => {
       this.sortCardsData = val;
       this.setSortByLabel();
-      let treasureHunt: TreasureHunt = this.treasureHuntService.treasureHunt.getValue();
-      let oppData = this.opportuntityCardsService.getOpportunityCardsData(treasureHunt, this.settings);
-      // this.setTeams(oppData);
-      this.setEquipments(oppData);
-      //this.setUtilityTypeOptions(oppData);
-      // this.setCalculatorOptions(oppData);
     });
-
 
     this.opportunityCardsSub = this.opportuntityCardsService.opportunityCards.subscribe(val => {
       this.opportunityCardsData = val;
@@ -98,11 +83,6 @@ export class TreasureChestMenuComponent implements OnInit {
     this.getNavbarWidth();
   }
 
-
-  toggleEquipment() {
-    this.displayEquipment = !this.displayEquipment;
-  }
-
   toggleAdditionalFilters() {
     if (this.displayAdditionalFiltersDropdown == 'hide') {
       this.displayAdditionalFiltersDropdown = 'show';
@@ -125,66 +105,6 @@ export class TreasureChestMenuComponent implements OnInit {
     this.treasureChestMenuService.deselectAll.next(false);
   }
 
-  setEquipments(oppData: Array<OpportunityCardData>) {
-    let sortByCpy: SortCardsData = JSON.parse(JSON.stringify(this.sortCardsData));
-    sortByCpy.equipments = [];
-    let sortedOppDataCpy: Array<OpportunityCardData> = this.sortCardsService.sortCards(JSON.parse(JSON.stringify(oppData)), sortByCpy);
-    let allEquipmentNames: Array<string> = this.treasureChestMenuService.getAllEquipment(sortedOppDataCpy);
-    this.equipments = new Array();
-    allEquipmentNames.forEach(equipment => {
-      let equipmentVal: { value: string, display: string } = processEquipmentOptions.find(option => { return option.value == equipment });
-      if (equipmentVal) {
-        this.equipments.push({
-          display: equipmentVal.display,
-          value: equipmentVal.value,
-          selected: this.sortCardsData.equipments.find(equipment => { return equipmentVal.value == equipment.value }) != undefined,
-          numCalcs: this.getFilteredCalcsByEquipment(sortedOppDataCpy, equipmentVal.value).length
-        });
-      }
-    });
-    let checkIsSelected: boolean = this.sortCardsData.equipments.length == 0;
-    this.equipments.unshift({ display: 'All', value: 'All', numCalcs: sortedOppDataCpy.length, selected: checkIsSelected });
-  }
-
-  setSelectedEquipment(option: FilterOption) {
-    let selectedFilters = this.getSelectedOptions(option, this.equipments)
-    this.sortCardsData.equipments = selectedFilters;
-    this.treasureChestMenuService.sortBy.next(this.sortCardsData);
-  }
-
-  //pass clicked option (selectedOption) and corresponding list to get all selected options
-  getSelectedOptions(selectedOption: FilterOption, optionList: Array<FilterOption>): Array<{ display: string, value: string }> {
-    let selected: Array<{ display: string, value: string }> = new Array();
-    //if the selection isn't all
-    if (selectedOption.value != 'All') {
-      selectedOption.selected = !selectedOption.selected;
-      optionList.forEach(option => {
-        //set All to false if an option is selected
-        if (selectedOption.selected && option.value == 'All') {
-          option.selected = false;
-        }
-        //add every selected option
-        if (option.selected == true) {
-          selected.push({ display: option.display, value: option.value });
-        }
-      });
-      //if none selected select all
-      if (selected.length == 0) {
-        let allOption: FilterOption = optionList.find(option => { return option.value == 'All' });
-        allOption.selected = true;
-      }
-    } else {
-      optionList.forEach(option => {
-        if (option.value == 'All') {
-          option.selected = true;
-        } else {
-          option.selected = false
-        }
-      });
-    }
-    return selected;
-  }
-
   // removeTeam(teamName: string, index: number) {
   //   this.sortCardsData.teams.splice(index, 1);
   //   this.teams.forEach(team => {
@@ -199,19 +119,19 @@ export class TreasureChestMenuComponent implements OnInit {
   //   this.treasureChestMenuService.sortBy.next(this.sortCardsData);
   // }
 
-  removeEquipment(equipmentItem: { display: string, value: string }, index: number) {
-    this.sortCardsData.equipments.splice(index, 1);
-    this.equipments.forEach(equipment => {
-      if (equipment.value == equipmentItem.value) {
-        equipment.selected = false;
-      }
-    });
-    if (this.sortCardsData.equipments.length == 0) {
-      let allOption: FilterOption = this.equipments.find(option => { return option.value == 'All' });
-      allOption.selected = true;
-    }
-    this.treasureChestMenuService.sortBy.next(this.sortCardsData);
-  }
+  // removeEquipment(equipmentItem: { display: string, value: string }, index: number) {
+  //   this.sortCardsData.equipments.splice(index, 1);
+  //   this.equipments.forEach(equipment => {
+  //     if (equipment.value == equipmentItem.value) {
+  //       equipment.selected = false;
+  //     }
+  //   });
+  //   if (this.sortCardsData.equipments.length == 0) {
+  //     let allOption: FilterOption = this.equipments.find(option => { return option.value == 'All' });
+  //     allOption.selected = true;
+  //   }
+  //   this.treasureChestMenuService.sortBy.next(this.sortCardsData);
+  // }
 
   // removeCalculator(calculatorItem: { display: string, value: string }, index: number) {
   //   this.sortCardsData.calculatorTypes.splice(index, 1);
@@ -275,13 +195,13 @@ export class TreasureChestMenuComponent implements OnInit {
     //     team.selected = true;
     //   }
     // })
-    this.equipments.forEach(equipment => {
-      if (equipment.value != 'All') {
-        equipment.selected = false;
-      } else {
-        equipment.selected = true;
-      }
-    })
+    // this.equipments.forEach(equipment => {
+    //   if (equipment.value != 'All') {
+    //     equipment.selected = false;
+    //   } else {
+    //     equipment.selected = true;
+    //   }
+    // })
     // this.calculatorTypeOptions.forEach(option => {
     //   if (option.value != 'All') {
     //     option.selected = false;
@@ -306,56 +226,6 @@ export class TreasureChestMenuComponent implements OnInit {
         this.navbarWidth = this.navbar.nativeElement.clientWidth * .95;
       }, 100);
     }
-  }
-
-  // setSelectedCalculator(calcOption: FilterOption) {
-  //   let selectedFilters = this.getSelectedOptions(calcOption, this.calculatorTypeOptions);
-  //   this.sortCardsData.calculatorTypes = selectedFilters;
-  //   this.treasureChestMenuService.sortBy.next(this.sortCardsData);
-  // }
-
-  // setCalculatorOptions(oppData: Array<OpportunityCardData>) {
-  //   let sortByCpy: SortCardsData = JSON.parse(JSON.stringify(this.sortCardsData));
-  //   sortByCpy.calculatorTypes = [];
-  //   let sortedOppDataCpy: Array<OpportunityCardData> = this.sortCardsService.sortCards(JSON.parse(JSON.stringify(oppData)), sortByCpy);
-  //   this.calculatorTypeOptions = new Array();
-  //   this.addCalculatorOption({ display: 'Lighting Replacement', value: 'lighting-replacement' }, sortedOppDataCpy, this.calculatorTypeOptions);
-  //   this.addCalculatorOption({ display: 'Opportunity Sheet', value: 'opportunity-sheet' }, sortedOppDataCpy, this.calculatorTypeOptions);
-  //   this.addCalculatorOption({ display: 'Replace Existing Motor', value: 'replace-existing' }, sortedOppDataCpy, this.calculatorTypeOptions);
-  //   this.addCalculatorOption({ display: 'Motor Drive', value: 'motor-drive' }, sortedOppDataCpy, this.calculatorTypeOptions);
-  //   this.addCalculatorOption({ display: 'Natural Gas Reduction', value: 'natural-gas-reduction' }, sortedOppDataCpy, this.calculatorTypeOptions);
-  //   this.addCalculatorOption({ display: 'Electricity Reduction', value: 'electricity-reduction' }, sortedOppDataCpy, this.calculatorTypeOptions);
-  //   this.addCalculatorOption({ display: 'Compressed Air Reduction', value: 'compressed-air-reduction' }, sortedOppDataCpy, this.calculatorTypeOptions);
-  //   this.addCalculatorOption({ display: 'Compressed Air Pressure Reduction', value: 'compressed-air-pressure-reduction' }, sortedOppDataCpy, this.calculatorTypeOptions);
-  //   this.addCalculatorOption({ display: 'Water Reduction', value: 'water-reduction' }, sortedOppDataCpy, this.calculatorTypeOptions);
-  //   this.addCalculatorOption({ display: 'Steam Reduction', value: 'steam-reduction' }, sortedOppDataCpy, this.calculatorTypeOptions);
-  //   this.addCalculatorOption({ display: 'Pipe Insulation Reduction', value: 'pipe-insulation-reduction' }, sortedOppDataCpy, this.calculatorTypeOptions);
-  //   let checkIsSelected: boolean = this.sortCardsData.calculatorTypes.length == 0;
-  //   this.calculatorTypeOptions.unshift({ display: 'All', value: 'All', numCalcs: sortedOppDataCpy.length, selected: checkIsSelected });
-  // }
-
-  // addCalculatorOption(calcOption: { display: string, value: string }, oppData: Array<OpportunityCardData>, calculatorTypeOptions: Array<FilterOption>) {
-  //   let numCalcs: number = this.getFilteredCalcsByCalculator(oppData, calcOption.value).length;
-  //   if (numCalcs != 0) {
-  //     let checkIsSelected: { display: string, value: string } = this.sortCardsData.calculatorTypes.find(calculatorType => { return calculatorType.value == calcOption.value });
-  //     calculatorTypeOptions.push({ display: calcOption.display, value: calcOption.value, numCalcs: numCalcs, selected: checkIsSelected != undefined });
-  //   }
-  // }
-
-  // getFilteredCalcsByCalculator(oppData: Array<OpportunityCardData>, calculatorType: string): Array<OpportunityCardData> {
-  //   let filteredCards: Array<OpportunityCardData> = _.filter(oppData, (data) => { return _.includes(data.opportunityType, calculatorType) });
-  //   return filteredCards;
-  // }
-
-  getFilteredCalcsByEquipment(oppData: Array<OpportunityCardData>, equipment: string): Array<OpportunityCardData> {
-    let filteredCards: Array<OpportunityCardData> = _.filter(oppData, (item: OpportunityCardData) => {
-      if (item.opportunitySheet) {
-        return _.includes(item.opportunitySheet.equipment, equipment);
-      } else {
-        return false;
-      }
-    });
-    return filteredCards;
   }
 
   openImportModal() {
